Extract resultscontrollers to a separate class
Add some headerdocs
